Desglozan recetas subidas, completadas y favoritas, además de poder verse cada receta

+++ app/src/main/java/Receta.kt
@@ -2,6 +2,6 @@ package teamfood.menufoodapp
 
 data class Receta  (var nombre: String,
                     var imagen: Int,
-                    var dificultad: Int,
+                    var dificultad: Float,
                     var ingredientes: ArrayList<Int>,
                     var pasos: String)
